requirements.txt
google-genai
python-dotenv
requests
beautifulsoup4
tqdm
